backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
clickhouse-connect
lark
openai
